django_backend/apps/notifications/urls.py
"""
URL configuration for notifications app.
"""
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from apps.notifications.views import NotificationViewSet, NotificationPreferenceViewSet

router = DefaultRouter()
router.register(r'notifications', NotificationViewSet, basename='notification')

urlpatterns = [
    path('', include(router.urls)),
    path('notifications/preferences/', 
         NotificationPreferenceViewSet.as_view({'get': 'list', 'post': 'create'}),
         name='notification-preferences'),
]
