django_backend/requirements/base.txt
# Django Escrow Platform - Base Requirements

# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9

# Environment
python-dotenv>=1.0.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# Security
argon2-cffi>=23.1.0
pyotp>=2.9.0

# Utilities
pillow>=10.0.0

# Nested routers for DRF
drf-nested-routers>=0.93.0
